connections_plus/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
Werkzeug==2.3.7
Jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
gunicorn==21.2.0 